a2ui/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "freudagent-a2ui"
version = "0.1.0"
description = "A2UI visual surfaces for the FreudAgent experiment harness"
requires-python = ">=3.10"
dependencies = [
    "freud-schema",
    "mcp>=1.0",
    "orjson>=3.9",
    "anyio>=4.5",
    "starlette>=0.36",
    "uvicorn>=0.25",
]

[project.optional-dependencies]
a2ui = ["a2ui-agent"]
anthropic = ["anthropic>=0.40"]
gemini = ["google-genai>=1.0"]
dev = ["pytest>=7.0", "httpx>=0.27.0"]

[tool.uv.sources]
freud-schema = { path = "../" }

[tool.setuptools]
py-modules = ["server", "bridge", "queries", "adapter", "providers", "prompt"]

[tool.pytest.ini_options]
pythonpath = ["."]
